fix warnings on x64 build. Xen apparently limits PFNs to 32 bits, so make this limitation stand out a little more

+++ xenhide/xenhide.c
@@ -51,7 +51,7 @@ DriverEntry(PDRIVER_OBJECT DriverObject, PUNICODE_STRING RegistryPath)
   ULONG BufLen = 200;
   PKEY_VALUE_PARTIAL_INFORMATION KeyPartialValue;
   int State = 0;
-  int StartPos = 0;
+  size_t StartPos = 0;
   WCHAR *SystemStartOptions;
   size_t SystemStartOptionsLen;
   size_t i;
